Move role related functions into its own file
